Apply _test.py suffix to test files
