+++ cognee/tests/unit/documents/PdfDocument_test.py
@@ -1,7 +1,8 @@
 import os
 import uuid
 
-from cognee.modules.data.processing.document_types.PdfDocument import PdfDocument
+from cognee.modules.data.processing.document_types.PdfDocument import \
+    PdfDocument
 
 GROUND_TRUTH = [
     {"word_count": 879, "len_text": 5622, "cut_type": "sentence_end"},
